services/profit/requirements.txt
firebase-admin==6.4.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.10.6
python-dotenv==1.1.1 
psycopg[binary]==3.2.3